Writed javadoc for AnswerServiceJpaImpl,ClarificationServiceJpaImpl,ChapterItemServiceJpaImpl. Changed ChapterEntity toString() method.Added AnswerServiceJpaImpl unit tests, ChapterItemServiceJpaImpl unit tests, ClarificationServiceJpaImpl unit tests.

+++ src/main/java/am/aca/service/ClarificationServiceJpaImpl.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.Objects;
 import java.util.Optional;
+
 @Service
 public class ClarificationServiceJpaImpl implements ClarificationService {
     private static final Logger log = LogManager.getLogger(ClarificationServiceJpaImpl.class);
@@ -18,18 +19,43 @@ public class ClarificationServiceJpaImpl implements ClarificationService {
         this.clarificationRepo = clarificationRepo;
     }
 
+    /**
+     * Save ClarificationEntity to DB.
+     *
+     * @param clarificationEntity <b>NonNull</b> with not empty clarificationText
+     * @return saved ClarificationEntity
+     * @throws NullPointerException     when argument is null
+     * @throws IllegalArgumentException when clarificationText is empty
+     */
     @Override
     public ClarificationEntity save(ClarificationEntity clarificationEntity) {
         log.debug("saving clarification " + clarificationEntity);
+        if (Objects.requireNonNull(clarificationEntity).getClarificationText().isEmpty()) {
+            IllegalArgumentException exception = new IllegalArgumentException(
+                    "Clarification text must be not empty " + clarificationEntity);
+            log.error(exception);
+            throw exception;
+        }
         ClarificationEntity save = clarificationRepo.save(clarificationEntity);
         log.debug("clarification saved " + save);
         return save;
     }
 
+    /**
+     * Find Clarification by id
+     *
+     * @param id must be > 0
+     * @return Optional<ClarificationEntity>
+     * @throws IllegalArgumentException when id < 1.
+     */
     @Override
     public Optional<ClarificationEntity> findOneById(Integer id) {
         log.debug("Searching Clarification by id " + id);
-
+        if (id < 1) {
+            IllegalArgumentException exception = new IllegalArgumentException("id must be > 0");
+            log.error(exception);
+            throw exception;
+        }
         Optional<ClarificationEntity> byId = clarificationRepo.findById(id);
         if (byId.isPresent()) {
             log.debug("Clarification found. " + byId.get());
@@ -39,26 +65,67 @@ public class ClarificationServiceJpaImpl implements ClarificationService {
         return byId;
     }
 
+    /**
+     * change text of Clarification.
+     *
+     * @param clarificationEntity <b>NonNull</b> with legal id (lid > 0)
+     * @param text                <b>NonNull</b> not empty String
+     * @return saved ClarificationEntity
+     * @throws NullPointerException     when one of the arguments is null
+     * @throws IllegalArgumentException when clarification id < 1 or text is empty
+     */
     @Override
     public ClarificationEntity changeText(ClarificationEntity clarificationEntity, String text) {
         log.debug("Changing " + clarificationEntity + "Text to " + text);
+        if (Objects.requireNonNull(clarificationEntity).getClarificationId() < 1 |
+                Objects.requireNonNull(text).isEmpty()) {
+            IllegalArgumentException exception = new IllegalArgumentException(clarificationEntity + " " + text);
+            log.error(exception);
+            throw exception;
+        }
         clarificationEntity.setClarificationText(text);
         ClarificationEntity save = clarificationRepo.save(clarificationEntity);
         log.debug("Clarification changed" + save);
         return save;
     }
 
+    /**
+     * Delete Clarification by id
+     *
+     * @param id must be > 0
+     * @return true, if successfully deleted
+     * @throws IllegalArgumentException when id < 1
+     */
     @Override
     public boolean deleteById(Integer id) {
         log.debug("Delete clarification by id");
+        if (id < 1) {
+            IllegalArgumentException exception = new IllegalArgumentException("id must be > 0 || id  = " + id);
+            log.error(exception);
+            throw exception;
+        }
         clarificationRepo.deleteById(id);
-        log.debug("Clarification with id " + id+" deleted.");
+        log.debug("Clarification with id " + id + " deleted.");
         return true;
     }
 
+    /**
+     * Delete clarification from DB.
+     *
+     * @param clarificationEntity <b>NonNull</b>
+     * @return true, if successfully deleted.
+     * @throws NullPointerException     when argument is null.
+     * @throws IllegalArgumentException when id < 1
+     */
     @Override
     public boolean delete(ClarificationEntity clarificationEntity) {
         log.debug("Deleting clarification" + clarificationEntity);
+        if (Objects.requireNonNull(clarificationEntity).getClarificationId() < 1) {
+            IllegalArgumentException exception =
+                    new IllegalArgumentException(Objects.toString(clarificationEntity));
+            log.error(exception);
+            throw exception;
+        }
         clarificationRepo.delete(clarificationEntity);
         log.debug("Deleted clarification" + clarificationEntity);
         return true;
